Make authentication tester class
